+++ lib/generators/footprinted/templates/footprinted.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+
+Footprinted.configure do |config|
+  # Configure the IP lookup service (default is :trackdown)
+  # config.ip_lookup_service = :trackdown
+end