+++ CHANGELOG.md
@@ -1,5 +1,10 @@
-## [Unreleased]
+# Changelog
 
 ## [0.1.0] - 2024-09-25
 
 - Initial release
+- Added Trackable concern for easy activity tracking
+- Integrated with trackdown gem for IP geolocation
+- Added customizable tracking associations
+- Created install generator for easy setup
+- Added configuration options